qa_chain.py
from langchain_openai import ChatOpenAI
from langchain.chains import create_retrieval_chain
from langchain.chains.combine_documents import create_stuff_documents_chain
from langchain_core.prompts import ChatPromptTemplate

def create_qa_chain(retriever):
    """
    Create a question-answering chain.
    """
    # Implement QA chain creation using ChatOpenAI, create_retrieval_chain, and create_stuff_documents_chain
    pass
