requirements.txt
langchain==0.2.0
langchain_openai==0.0.5
langchain_community==0.0.20
python-dotenv==1.0.1
chromadb==0.4.22
